backend/calendar-sync/requirements.txt
psycopg2-binary==2.9.9
google-auth==2.26.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0
